gui/requirements.txt
PyYAML
tkinter
